huge code base cleanup
