+++ src/services/db/types/auth.ts
@@ -1,0 +1,26 @@
+// === DISCORD TYPES ===
+
+export interface DiscordUser {
+  id: string;
+  username: string;
+  discriminator: string;
+  avatar?: string;
+  email?: string;
+}
+
+// === AUTH TYPES ===
+
+import type { User } from './user';
+
+export interface AuthResult {
+  success: boolean;
+  user?: User;
+  error?: string;
+}
+
+export interface AuthData {
+  accessToken: string;
+  tokenType: string;
+  user: User;
+  loginTime: string;
+}
