+++ src/services/ocr/templateManager.ts
@@ -1,59 +1,15 @@
 import fs from 'fs';
 import path from 'path';
-import { UserPreferences } from '../db/types';
-
-export interface ItemTemplate {
-    name: string;
-    variations?: string[];
-    fuzzyMatch?: boolean;
-    fuzzyPattern?: string;
-    minSimilarity?: number;
-    value?: number;
-}
-
-export interface CropArea {
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-}
-
-export interface TextFilters {
-    minConfidence?: number;
-    allowedCharacters?: string;
-    minLength?: number;
-}
-
-export interface Pattern {
-    type: string;
-    regex: string;
-    groups: string[];
-}
-
-export interface LocationTemplateData {
-    name: string;
-    cropArea?: CropArea;
-    textFilters?: TextFilters;
-    items: ItemTemplate[];
-    patterns?: Pattern[];
-}
-
-export interface ItemMatcher {
-    exact: Set<string>;
-    fuzzy: Array<{
-        name: string;
-        pattern: RegExp;
-        minSimilarity: number;
-    }>;
-}
-
-export interface MatchResult {
-    match: boolean;
-    item?: string;
-    confidence?: number;
-    method?: 'exact' | 'fuzzy';
-    text?: string;
-}
+import { UserPreferences } from '../db/types/user';
+import { 
+    ItemTemplate, 
+    CropArea, 
+    TextFilters, 
+    Pattern, 
+    LocationTemplateData, 
+    ItemMatcher, 
+    MatchResult 
+} from '../db/types/ocr';
 
 export class LocationTemplate {
     public name: string;
